+++ nextgisweb/tileset/nodepkg/resource-widget/index.js
@@ -1,0 +1,3 @@
+/** @entrypoint */
+export { Widget as widget } from "./Widget";
+export { Store as store } from "./Store";