+++ nextgisweb/tileset/view.py
@@ -1,0 +1,13 @@
+from nextgisweb.resource import Widget
+
+from .model import Tileset
+
+
+class TilesetWidget(Widget):
+    resource = Tileset
+    operation = ('create', 'update')
+    amdmod = '@nextgisweb/tileset/resource-widget'
+
+
+def setup_pyramid(comp, config):
+    pass